we do a little refactoring
as a treat
